test(cli): split bundle option tests
